requirements.txt
cryptography
numpy
pandas
python-dotenv
requests
scikit-learn
